refactor: rename token-creator to token-generator and update related references
