+++ backend/go.work
@@ -2,6 +2,6 @@ go 1.24.1
 
 use (
 	./app/api-service
-	./app/token-creator
+	./app/token-generator
 	./core
 )
